+++ nautilus_trader/common/factories.pyx
@@ -22,6 +22,7 @@ from nautilus_trader.core.correctness cimport Condition
 from nautilus_trader.model.c_enums.contingency_type cimport ContingencyType
 from nautilus_trader.model.c_enums.order_side cimport OrderSide
 from nautilus_trader.model.c_enums.time_in_force cimport TimeInForce
+from nautilus_trader.model.c_enums.trailing_offset_type cimport TrailingOffsetType
 from nautilus_trader.model.identifiers cimport ClientOrderId
 from nautilus_trader.model.identifiers cimport InstrumentId
 from nautilus_trader.model.identifiers cimport OrderListId
@@ -32,6 +33,8 @@ from nautilus_trader.model.orders.base cimport Order
 from nautilus_trader.model.orders.limit cimport LimitOrder
 from nautilus_trader.model.orders.list cimport OrderList
 from nautilus_trader.model.orders.stop_market cimport StopMarketOrder
+from nautilus_trader.model.orders.trailing_stop_limit cimport TrailingStopLimitOrder
+from nautilus_trader.model.orders.trailing_stop_market cimport TrailingStopMarketOrder
 
 
 cdef class OrderFactory:
@@ -260,14 +263,14 @@ cdef class OrderFactory:
         OrderSide order_side,
         Quantity quantity,
         Price trigger_price,
-        TriggerMethod trigger=TriggerMethod.DEFAULT,
+        TriggerType trigger_type=TriggerType.DEFAULT,
         TimeInForce time_in_force=TimeInForce.GTC,
         datetime expiration=None,
         bint reduce_only=False,
         str tags=None,
     ):
         """
-        Create a new stop-market order.
+        Create a new stop-market trigger order.
 
         If the time-in-force is ``GTD`` then a valid expire time must be given.
 
@@ -281,8 +284,8 @@ cdef class OrderFactory:
             The orders quantity (> 0).
         trigger_price : Price
             The orders trigger price (STOP).
-        trigger : TriggerMethod
-            The order trigger method.
+        trigger_type : TriggerType
+            The order trigger type.
         time_in_force : TimeInForce, optional
             The orders time-in-force.
         expiration : datetime, optional
@@ -313,7 +316,7 @@ cdef class OrderFactory:
             order_side=order_side,
             quantity=quantity,
             trigger_price=trigger_price,
-            trigger=trigger,
+            trigger_type=trigger_type,
             time_in_force=time_in_force,
             expiration=expiration,
             init_id=self._uuid_factory.generate(),
@@ -334,7 +337,7 @@ cdef class OrderFactory:
         Quantity quantity,
         Price price,
         Price trigger_price,
-        TriggerMethod trigger=TriggerMethod.DEFAULT,
+        TriggerType trigger_type=TriggerType.DEFAULT,
         TimeInForce time_in_force=TimeInForce.GTC,
         datetime expiration=None,
         bint post_only=False,
@@ -343,7 +346,7 @@ cdef class OrderFactory:
         str tags=None,
     ):
         """
-        Create a new stop-limit order.
+        Create a new stop-limit trigger order.
 
         If the time-in-force is ``GTD`` then a valid expire time must be given.
 
@@ -359,8 +362,8 @@ cdef class OrderFactory:
             The orders limit price.
         trigger_price : Price
             The orders trigger stop price.
-        trigger : TriggerMethod
-            The order trigger method.
+        trigger_type : TriggerType
+            The order trigger type.
         time_in_force : TimeInForce, optional
             The orders time-in-force.
         expiration : datetime, optional
@@ -398,7 +401,187 @@ cdef class OrderFactory:
             quantity=quantity,
             price=price,
             trigger_price=trigger_price,
-            trigger=trigger,
+            trigger_type=trigger_type,
+            time_in_force=time_in_force,
+            expiration=expiration,
+            init_id=self._uuid_factory.generate(),
+            ts_init=self._clock.timestamp_ns(),
+            post_only=post_only,
+            reduce_only=reduce_only,
+            display_qty=display_qty,
+            order_list_id=None,
+            parent_order_id=None,
+            child_order_ids=None,
+            contingency=ContingencyType.NONE,
+            contingency_ids=None,
+            tags=tags,
+        )
+
+    cpdef TrailingStopMarketOrder trailing_stop_market(
+        self,
+        InstrumentId instrument_id,
+        OrderSide order_side,
+        Quantity quantity,
+        Price trigger_price,
+        trailing_offset: Decimal,
+        TriggerType trigger_type=TriggerType.DEFAULT,
+        TrailingOffsetType offset_type=TrailingOffsetType.PRICE,
+        TimeInForce time_in_force=TimeInForce.GTC,
+        datetime expiration=None,
+        bint reduce_only=False,
+        str tags=None,
+    ):
+        """
+        Create a new trailing stop-market trigger order.
+
+        If the time-in-force is ``GTD`` then a valid expire time must be given.
+
+        Parameters
+        ----------
+        instrument_id : InstrumentId
+            The orders instrument ID.
+        order_side : OrderSide {``BUY``, ``SELL``}
+            The orders side.
+        quantity : Quantity
+            The orders quantity (> 0).
+        trigger_price : Price
+            The orders trigger price (STOP).
+        trailing_offset : Decimal
+            The trailing offset for the trigger (STOP) price.
+        trigger_type : TriggerType, optional
+            The order trigger type.
+        offset_type : TrailingOffsetType, optional
+            The order trailing offset type.
+        time_in_force : TimeInForce, optional
+            The orders time-in-force.
+        expiration : datetime, optional
+            The order expiration (for ``GTD`` orders).
+        reduce_only : bool, optional
+            If the order carries the 'reduce-only' execution instruction.
+        tags : str, optional
+            The custom user tags for the order. These are optional and can
+            contain any arbitrary delimiter if required.
+
+        Returns
+        -------
+        TrailingStopMarketOrder
+
+        Raises
+        ------
+        ValueError
+            If `quantity` is not positive (> 0).
+        ValueError
+            If `time_in_force` is ``GTD`` and `expiration` is ``None`` or <= UNIX epoch.
+
+        """
+        return TrailingStopMarketOrder(
+            trader_id=self.trader_id,
+            strategy_id=self.strategy_id,
+            instrument_id=instrument_id,
+            client_order_id=self._id_generator.generate(),
+            order_side=order_side,
+            quantity=quantity,
+            trigger_price=trigger_price,
+            trigger_type=trigger_type,
+            trailing_offset=trailing_offset,
+            offset_type=offset_type,
+            time_in_force=time_in_force,
+            expiration=expiration,
+            init_id=self._uuid_factory.generate(),
+            ts_init=self._clock.timestamp_ns(),
+            reduce_only=reduce_only,
+            order_list_id=None,
+            parent_order_id=None,
+            child_order_ids=None,
+            contingency=ContingencyType.NONE,
+            contingency_ids=None,
+            tags=tags,
+        )
+
+    cpdef TrailingStopLimitOrder trailing_stop_limit(
+        self,
+        InstrumentId instrument_id,
+        OrderSide order_side,
+        Quantity quantity,
+        Price price,
+        Price trigger_price,
+        limit_offset: Decimal,
+        trailing_offset: Decimal,
+        TriggerType trigger_type=TriggerType.DEFAULT,
+        TrailingOffsetType offset_type=TrailingOffsetType.PRICE,
+        TimeInForce time_in_force=TimeInForce.GTC,
+        datetime expiration=None,
+        bint post_only=False,
+        bint reduce_only=False,
+        Quantity display_qty=None,
+        str tags=None,
+    ):
+        """
+        Create a new trailing stop-limit trigger order.
+
+        If the time-in-force is ``GTD`` then a valid expire time must be given.
+
+        Parameters
+        ----------
+        instrument_id : InstrumentId
+            The orders instrument ID.
+        order_side : OrderSide {``BUY``, ``SELL``}
+            The orders side.
+        quantity : Quantity
+            The orders quantity (> 0).
+        price : Price
+            The orders limit price.
+        trigger_price : Price
+            The orders trigger stop price.
+        trailing_offset : Decimal
+            The trailing offset for the trigger (STOP) price.
+        limit_offset : Decimal
+            The trailing offset for the order (LIMIT) price.
+        trigger_type : TriggerType, optional
+            The order trigger type.
+        offset_type : TrailingOffsetType, optional
+            The order trailing offset type.
+        time_in_force : TimeInForce, optional
+            The orders time-in-force.
+        expiration : datetime, optional
+            The order expiration (for ``GTD`` orders).
+        post_only : bool, optional
+            If the order will only provide liquidity (make a market).
+        reduce_only : bool, optional
+            If the order carries the 'reduce-only' execution instruction.
+        display_qty : Quantity, optional
+            The quantity of the order to display on the public book (iceberg).
+        tags : str, optional
+            The custom user tags for the order. These are optional and can
+            contain any arbitrary delimiter if required.
+
+        Returns
+        -------
+        TrailingStopLimitOrder
+
+        Raises
+        ------
+        ValueError
+            If `quantity` is not positive (> 0).
+        ValueError
+            If `time_in_force` is ``GTD`` and `expiration` is ``None`` or <= UNIX epoch.
+        ValueError
+            If `display_qty` is negative (< 0) or greater than `quantity`.
+
+        """
+        return TrailingStopLimitOrder(
+            trader_id=self.trader_id,
+            strategy_id=self.strategy_id,
+            instrument_id=instrument_id,
+            client_order_id=self._id_generator.generate(),
+            order_side=order_side,
+            quantity=quantity,
+            price=price,
+            trigger_price=trigger_price,
+            trigger_type=trigger_type,
+            limit_offset=limit_offset,
+            trailing_offset=trailing_offset,
+            offset_type=offset_type,
             time_in_force=time_in_force,
             expiration=expiration,
             init_id=self._uuid_factory.generate(),
@@ -435,9 +618,9 @@ cdef class OrderFactory:
         quantity : Quantity
             The entry orders quantity (> 0).
         stop_loss : Price
-            The stop-loss child order stop price.
+            The stop-loss child order trigger (STOP) price.
         take_profit : Price
-            The take-profit child order limit price.
+            The take-profit child order (LIMIT) price.
         tif_bracket : TimeInForce {``DAY``, ``GTC``}, optional
             The bracket orders time-in-force .
 
@@ -499,7 +682,7 @@ cdef class OrderFactory:
             order_side=Order.opposite_side_c(entry_order.side),
             quantity=quantity,
             trigger_price=stop_loss,
-            trigger=TriggerMethod.DEFAULT,
+            trigger_type=TriggerType.DEFAULT,
             time_in_force=tif_bracket,
             expiration=None,
             init_id=self._uuid_factory.generate(),
@@ -567,9 +750,9 @@ cdef class OrderFactory:
         entry : Price
             The entry LIMIT order price.
         stop_loss : Price
-            The stop-loss child order stop price.
+            The stop-loss child order trigger (STOP) price.
         take_profit : Price
-            The take-profit child order limit price.
+            The take-profit child order (LIMIT) price.
         tif : TimeInForce {``DAY``, ``GTC``}, optional
             The entry orders time-in-force .
         expiration : datetime, optional
@@ -646,7 +829,7 @@ cdef class OrderFactory:
             order_side=Order.opposite_side_c(entry_order.side),
             quantity=quantity,
             trigger_price=stop_loss,
-            trigger=TriggerMethod.DEFAULT,
+            trigger_type=TriggerType.DEFAULT,
             time_in_force=tif_bracket,
             expiration=None,
             init_id=self._uuid_factory.generate(),