+++ nautilus_trader/model/orders/trailing_stop_market.pyx
@@ -13,6 +13,8 @@
 #  limitations under the License.
 # -------------------------------------------------------------------------------------------------
 
+from decimal import Decimal
+
 from cpython.datetime cimport datetime
 from libc.stdint cimport int64_t
 
@@ -30,10 +32,10 @@ from nautilus_trader.model.c_enums.order_type cimport OrderType
 from nautilus_trader.model.c_enums.order_type cimport OrderTypeParser
 from nautilus_trader.model.c_enums.time_in_force cimport TimeInForce
 from nautilus_trader.model.c_enums.time_in_force cimport TimeInForceParser
+from nautilus_trader.model.c_enums.trailing_offset_type cimport TrailingOffsetTypeParser
 from nautilus_trader.model.c_enums.trigger_type cimport TriggerType
 from nautilus_trader.model.c_enums.trigger_type cimport TriggerTypeParser
 from nautilus_trader.model.events.order cimport OrderInitialized
-from nautilus_trader.model.events.order cimport OrderTriggered
 from nautilus_trader.model.events.order cimport OrderUpdated
 from nautilus_trader.model.identifiers cimport ClientOrderId
 from nautilus_trader.model.identifiers cimport InstrumentId
@@ -45,21 +47,19 @@ from nautilus_trader.model.objects cimport Quantity
 from nautilus_trader.model.orders.base cimport Order
 
 
-cdef class StopLimitOrder(Order):
+cdef class TrailingStopMarketOrder(Order):
     """
-    Represents a stop-limit trigger order.
+    Represents a trailing stop-market trigger order.
 
-    A stop-limit order is an instruction to submit a buy or sell limit order
-    when the user-specified stop trigger price is attained or penetrated. The
-    order has two basic components: the stop price and the limit price. When a
-    trade has occurred at or through the stop price, the order becomes
-    executable and enters the market as a limit order, which is an order to buy
-    or sell at a specified price or better.
-
-    A stop-limit eliminates the price risk associated with a stop order where
-    the execution price cannot be guaranteed, but exposes the trader to the
-    risk that the order may never fill even if the stop price is reached. The
-    trader could "miss the market" altogether.
+    A sell trailing stop order sets the stop price at a fixed amount below the
+    market price with an attached "trailing" offset. As the market price rises,
+    the stop price rises by the trail amount, but if the stock price falls, the
+    stop loss price doesn't change, and a market order is submitted when the
+    stop price is hit. This technique is designed to allow a trader to
+    specify a limit on the maximum possible loss, without setting a limit on
+    the maximum possible gain. "Buy" trailing stop orders are the mirror image
+    of sell trailing stop orders, and are most appropriate for use in falling
+    markets.
 
     Parameters
     ----------
@@ -75,12 +75,14 @@ cdef class StopLimitOrder(Order):
         The order side.
     quantity : Quantity
         The order quantity (> 0).
-    price : Price
-        The order price (LIMIT).
     trigger_price : Price
         The order trigger price (STOP).
     trigger_type : TriggerType
         The order trigger type.
+    trailing_offset : Decimal
+        The trailing offset for the trigger (STOP) price.
+    offset_type : TrailingOffsetType
+        The order trailing offset type.
     time_in_force : TimeInForce
         The order time-in-force.
     expiration : datetime, optional
@@ -89,12 +91,8 @@ cdef class StopLimitOrder(Order):
         The order initialization event ID.
     ts_init : int64
         The UNIX timestamp (nanoseconds) when the object was initialized.
-    post_only : bool, optional
-        If the ``LIMIT`` order will only provide liquidity (once triggered).
     reduce_only : bool, optional
-        If the ``LIMIT`` order carries the 'reduce-only' execution instruction.
-    display_qty : Quantity, optional
-        The quantity of the ``LIMIT`` order to display on the public book (iceberg).
+        If the order carries the 'reduce-only' execution instruction.
     order_list_id : OrderListId, optional
         The order list ID associated with the order.
     parent_order_id : ClientOrderId, optional
@@ -115,10 +113,7 @@ cdef class StopLimitOrder(Order):
         If `quantity` is not positive (> 0).
     ValueError
         If `time_in_force` is ``GTD`` and `expiration` is ``None`` or <= UNIX epoch.
-    ValueError
-        If `display_qty` is negative (< 0) or greater than `quantity`.
     """
-
     def __init__(
         self,
         TraderId trader_id not None,
@@ -127,16 +122,15 @@ cdef class StopLimitOrder(Order):
         ClientOrderId client_order_id not None,
         OrderSide order_side,
         Quantity quantity not None,
-        Price price not None,
         Price trigger_price not None,
         TriggerType trigger_type,
+        trailing_offset: Decimal,
+        TrailingOffsetType offset_type,
         TimeInForce time_in_force,
         datetime expiration,  # Can be None
         UUID4 init_id not None,
         int64_t ts_init,
-        bint post_only=False,
         bint reduce_only=False,
-        Quantity display_qty=None,
         OrderListId order_list_id=None,
         ClientOrderId parent_order_id=None,
         list child_order_ids=None,
@@ -153,18 +147,14 @@ cdef class StopLimitOrder(Order):
         else:
             # Should not have an expire time
             Condition.none(expiration, "expiration")
-        Condition.true(
-            display_qty is None or 0 <= display_qty <= quantity,
-            fail_msg="display_qty was negative or greater than order quantity",
-        )
 
         # Set options
         cdef dict options = {
-            "price": str(price),
             "trigger_price": str(trigger_price),
             "trigger_type": TriggerTypeParser.to_str(trigger_type),
+            "trailing_offset": str(trailing_offset),
+            "offset_type": TrailingOffsetTypeParser.to_str(offset_type),
             "expiration_ns": expiration_ns if expiration_ns > 0 else None,
-            "display_qty": str(display_qty) if display_qty is not None else None,
         }
 
         # Create initialization event
@@ -174,10 +164,10 @@ cdef class StopLimitOrder(Order):
             instrument_id=instrument_id,
             client_order_id=client_order_id,
             order_side=order_side,
-            order_type=OrderType.STOP_LIMIT,
+            order_type=OrderType.TRAILING_STOP_MARKET,
             quantity=quantity,
             time_in_force=time_in_force,
-            post_only=post_only,
+            post_only=False,
             reduce_only=reduce_only,
             options=options,
             order_list_id=order_list_id,
@@ -191,14 +181,12 @@ cdef class StopLimitOrder(Order):
         )
         super().__init__(init=init)
 
-        self.price = price
         self.trigger_price = trigger_price
         self.trigger_type = trigger_type
+        self.trailing_offset = trailing_offset
+        self.offset_type = offset_type
         self.expiration = expiration
         self.expiration_ns = expiration_ns
-        self.display_qty = display_qty
-        self.is_triggered = False
-        self.ts_triggered = 0
 
     cpdef str info(self):
         """
@@ -212,8 +200,9 @@ cdef class StopLimitOrder(Order):
         cdef str expiration_str = "" if self.expiration is None else f" {format_iso8601(self.expiration)}"
         return (
             f"{OrderSideParser.to_str(self.side)} {self.quantity.to_str()} {self.instrument_id} "
-            f"{OrderTypeParser.to_str(self.type)} @ {self.trigger_price}-STOP"
-            f"[{TriggerTypeParser.to_str(self.trigger_type)}] {self.price}-LIMIT "
+            f"{OrderTypeParser.to_str(self.type)} @ {self.trigger_price}"
+            f"[{TriggerTypeParser.to_str(self.trigger_type)}] "
+            f"{self.trailing_offset}-TRAILING_OFFSET[{TrailingOffsetTypeParser.to_str(self.offset_type)}] "
             f"{TimeInForceParser.to_str(self.time_in_force)}{expiration_str}"
         )
 
@@ -231,16 +220,17 @@ cdef class StopLimitOrder(Order):
             "strategy_id": self.strategy_id.value,
             "instrument_id": self.instrument_id.value,
             "client_order_id": self.client_order_id.value,
-            "venue_order_id": self.venue_order_id.value if self.venue_order_id else None,
-            "position_id": self.position_id if self.position_id else None,
+            "venue_order_id": self.venue_order_id if self.venue_order_id else None,
+            "position_id": self.position_id.value if self.position_id else None,
             "account_id": self.account_id.value if self.account_id else None,
             "last_trade_id": self.last_trade_id.value if self.last_trade_id else None,
             "type": OrderTypeParser.to_str(self.type),
             "side": OrderSideParser.to_str(self.side),
             "quantity": str(self.quantity),
-            "price": str(self.price),
             "trigger_price": str(self.trigger_price),
             "trigger_type": TriggerTypeParser.to_str(self.trigger_type),
+            "trailing_offset": str(self.trailing_offset),
+            "offset_type": TrailingOffsetTypeParser.to_str(self.offset_type),
             "expiration_ns": self.expiration_ns if self.expiration_ns > 0 else None,
             "time_in_force": TimeInForceParser.to_str(self.time_in_force),
             "filled_qty": str(self.filled_qty),
@@ -248,9 +238,7 @@ cdef class StopLimitOrder(Order):
             "avg_px": str(self.avg_px) if self.avg_px else None,
             "slippage": str(self.slippage),
             "status": self._fsm.state_string_c(),
-            "is_post_only": self.is_post_only,
             "is_reduce_only": self.is_reduce_only,
-            "display_qty": str(self.display_qty) if self.display_qty is not None else None,
             "order_list_id": self.order_list_id,
             "parent_order_id": self.parent_order_id,
             "child_order_ids": ",".join([o.value for o in self.child_order_ids]) if self.child_order_ids is not None else None,  # noqa
@@ -262,9 +250,9 @@ cdef class StopLimitOrder(Order):
         }
 
     @staticmethod
-    cdef StopLimitOrder create(OrderInitialized init):
+    cdef TrailingStopMarketOrder create(OrderInitialized init):
         """
-        Return a stop-limit order from the given initialized event.
+        Return a stop-market order from the given initialized event.
 
         Parameters
         ----------
@@ -273,36 +261,33 @@ cdef class StopLimitOrder(Order):
 
         Returns
         -------
-        StopLimitOrder
+        StopMarketOrder
 
         Raises
         ------
         ValueError
-            If `init.type` is not equal to ``STOP_LIMIT``.
+            If `init.type` is not equal to ``STOP_MARKET``.
 
         """
         Condition.not_none(init, "init")
-        Condition.equal(init.type, OrderType.STOP_LIMIT, "init.type", "OrderType")
+        Condition.equal(init.type, OrderType.TRAILING_STOP_MARKET, "init.type", "OrderType")
 
-        cdef str display_qty_str = init.options["display_qty"]
-
-        return StopLimitOrder(
+        return TrailingStopMarketOrder(
             trader_id=init.trader_id,
             strategy_id=init.strategy_id,
             instrument_id=init.instrument_id,
             client_order_id=init.client_order_id,
             order_side=init.side,
             quantity=init.quantity,
-            price=Price.from_str_c(init.options["price"]),
             trigger_price=Price.from_str_c(init.options["trigger_price"]),
             trigger_type=TriggerTypeParser.from_str(init.options["trigger_type"]),
+            trailing_offset=Decimal(init.options["trailing_offset"]),
+            offset_type=TrailingOffsetTypeParser.from_str(init.options["offset_type"]),
             time_in_force=init.time_in_force,
             expiration=maybe_unix_nanos_to_dt(init.options["expiration_ns"]),
             init_id=init.id,
             ts_init=init.ts_init,
-            post_only=init.post_only,
             reduce_only=init.reduce_only,
-            display_qty=Quantity.from_str_c(display_qty_str) if display_qty_str is not None else None,
             order_list_id=init.order_list_id,
             parent_order_id=init.parent_order_id,
             child_order_ids=init.child_order_ids,
@@ -318,17 +303,11 @@ cdef class StopLimitOrder(Order):
         if event.quantity is not None:
             self.quantity = event.quantity
             self.leaves_qty = Quantity(self.quantity - self.filled_qty, self.quantity.precision)
-        if event.price is not None:
-            self.price = event.price
         if event.trigger_price is not None:
             self.trigger_price = event.trigger_price
 
-    cdef void _triggered(self, OrderTriggered event) except *:
-        self.is_triggered = True
-        self.ts_triggered = event.ts_event
-
     cdef void _set_slippage(self) except *:
         if self.side == OrderSide.BUY:
-            self.slippage = self.avg_px - self.price
+            self.slippage = self.avg_px - self.trigger_price
         elif self.side == OrderSide.SELL:
-            self.slippage = self.price - self.avg_px
+            self.slippage = self.trigger_price - self.avg_px