+++ tests/unit_tests/execution/test_execution_reports.py
@@ -13,6 +13,8 @@
 #  limitations under the License.
 # -------------------------------------------------------------------------------------------------
 
+from decimal import Decimal
+
 from nautilus_trader.execution.reports import ExecutionMassStatus
 from nautilus_trader.execution.reports import OrderStatusReport
 from nautilus_trader.execution.reports import PositionStatusReport
@@ -25,7 +27,8 @@ from nautilus_trader.model.enums import OrderStatus
 from nautilus_trader.model.enums import OrderType
 from nautilus_trader.model.enums import PositionSide
 from nautilus_trader.model.enums import TimeInForce
-from nautilus_trader.model.enums import TriggerMethod
+from nautilus_trader.model.enums import TrailingOffsetType
+from nautilus_trader.model.enums import TriggerType
 from nautilus_trader.model.identifiers import ClientId
 from nautilus_trader.model.identifiers import ClientOrderId
 from nautilus_trader.model.identifiers import OrderListId
@@ -56,7 +59,10 @@ class TestExecutionReports:
             order_status=OrderStatus.REJECTED,
             price=Price.from_str("0.90090"),
             trigger_price=Price.from_str("0.90100"),
-            trigger=TriggerMethod.DEFAULT,
+            trigger_type=TriggerType.DEFAULT,
+            limit_offset=None,
+            trailing_offset=Decimal("0.00010"),
+            offset_type=TrailingOffsetType.PRICE,
             quantity=Quantity.from_int(1_000_000),
             filled_qty=Quantity.from_int(0),
             display_qty=None,
@@ -73,11 +79,11 @@ class TestExecutionReports:
         # Assert
         assert (
             str(report)
-            == "OrderStatusReport(client_order_id=O-123456, order_list_id=1, venue_order_id=2, order_side=SELL, order_type=STOP_LIMIT, contingency=OCO, time_in_force=DAY, order_status=REJECTED, price=0.90090, trigger_price=0.90100, trigger=DEFAULT, quantity=1000000, filled_qty=0, leaves_qty=1000000, display_qty=None, avg_px=None, post_only=True, reduce_only=False, reject_reason=SOME_REASON, ts_accepted=1000000, ts_triggered=1500000, ts_last=2000000, ts_init=3000000)"  # noqa
+            == "OrderStatusReport(client_order_id=O-123456, order_list_id=1, venue_order_id=2, order_side=SELL, order_type=STOP_LIMIT, contingency=OCO, time_in_force=DAY, order_status=REJECTED, price=0.90090, trigger_price=0.90100, trigger_type=DEFAULT, limit_offset=None, trailing_offset=0.00010, offset_type=PRICE, quantity=1000000, filled_qty=0, leaves_qty=1000000, display_qty=None, avg_px=None, post_only=True, reduce_only=False, reject_reason=SOME_REASON, ts_accepted=1000000, ts_triggered=1500000, ts_last=2000000, ts_init=3000000)"  # noqa
         )
         assert (
             repr(report)
-            == "OrderStatusReport(client_order_id=O-123456, order_list_id=1, venue_order_id=2, order_side=SELL, order_type=STOP_LIMIT, contingency=OCO, time_in_force=DAY, order_status=REJECTED, price=0.90090, trigger_price=0.90100, trigger=DEFAULT, quantity=1000000, filled_qty=0, leaves_qty=1000000, display_qty=None, avg_px=None, post_only=True, reduce_only=False, reject_reason=SOME_REASON, ts_accepted=1000000, ts_triggered=1500000, ts_last=2000000, ts_init=3000000)"  # noqa
+            == "OrderStatusReport(client_order_id=O-123456, order_list_id=1, venue_order_id=2, order_side=SELL, order_type=STOP_LIMIT, contingency=OCO, time_in_force=DAY, order_status=REJECTED, price=0.90090, trigger_price=0.90100, trigger_type=DEFAULT, limit_offset=None, trailing_offset=0.00010, offset_type=PRICE, quantity=1000000, filled_qty=0, leaves_qty=1000000, display_qty=None, avg_px=None, post_only=True, reduce_only=False, reject_reason=SOME_REASON, ts_accepted=1000000, ts_triggered=1500000, ts_last=2000000, ts_init=3000000)"  # noqa
         )
 
     def test_instantiate_trade_report(self):
@@ -176,7 +182,10 @@ class TestExecutionReports:
             order_status=OrderStatus.REJECTED,
             price=Price.from_str("0.90090"),
             trigger_price=Price.from_str("0.90100"),
-            trigger=TriggerMethod.DEFAULT,
+            trigger_type=TriggerType.DEFAULT,
+            limit_offset=None,
+            trailing_offset=Decimal("0.00010"),
+            offset_type=TrailingOffsetType.PRICE,
             quantity=Quantity.from_int(1_000_000),
             filled_qty=Quantity.from_int(0),
             display_qty=None,
@@ -197,11 +206,11 @@ class TestExecutionReports:
         assert mass_status.order_reports()[venue_order_id] == report
         assert (
             repr(mass_status)
-            == "ExecutionMassStatus(client_id=IB, account_id=SIM-000, order_reports={VenueOrderId('2'): OrderStatusReport(client_order_id=O-123456, order_list_id=1, venue_order_id=2, order_side=SELL, order_type=STOP_LIMIT, contingency=OCO, time_in_force=DAY, order_status=REJECTED, price=0.90090, trigger_price=0.90100, trigger=DEFAULT, quantity=1000000, filled_qty=0, leaves_qty=1000000, display_qty=None, avg_px=None, post_only=True, reduce_only=False, reject_reason=SOME_REASON, ts_accepted=1000000, ts_triggered=0, ts_last=2000000, ts_init=3000000)}, trade_reports={}, position_reports={}, ts_init=0)"  # noqa
+            == "ExecutionMassStatus(client_id=IB, account_id=SIM-000, order_reports={VenueOrderId('2'): OrderStatusReport(client_order_id=O-123456, order_list_id=1, venue_order_id=2, order_side=SELL, order_type=STOP_LIMIT, contingency=OCO, time_in_force=DAY, order_status=REJECTED, price=0.90090, trigger_price=0.90100, trigger_type=DEFAULT, limit_offset=None, trailing_offset=0.00010, offset_type=PRICE, quantity=1000000, filled_qty=0, leaves_qty=1000000, display_qty=None, avg_px=None, post_only=True, reduce_only=False, reject_reason=SOME_REASON, ts_accepted=1000000, ts_triggered=0, ts_last=2000000, ts_init=3000000)}, trade_reports={}, position_reports={}, ts_init=0)"  # noqa
         )
         assert (
             repr(report)
-            == "OrderStatusReport(client_order_id=O-123456, order_list_id=1, venue_order_id=2, order_side=SELL, order_type=STOP_LIMIT, contingency=OCO, time_in_force=DAY, order_status=REJECTED, price=0.90090, trigger_price=0.90100, trigger=DEFAULT, quantity=1000000, filled_qty=0, leaves_qty=1000000, display_qty=None, avg_px=None, post_only=True, reduce_only=False, reject_reason=SOME_REASON, ts_accepted=1000000, ts_triggered=0, ts_last=2000000, ts_init=3000000)"  # noqa
+            == "OrderStatusReport(client_order_id=O-123456, order_list_id=1, venue_order_id=2, order_side=SELL, order_type=STOP_LIMIT, contingency=OCO, time_in_force=DAY, order_status=REJECTED, price=0.90090, trigger_price=0.90100, trigger_type=DEFAULT, limit_offset=None, trailing_offset=0.00010, offset_type=PRICE, quantity=1000000, filled_qty=0, leaves_qty=1000000, display_qty=None, avg_px=None, post_only=True, reduce_only=False, reject_reason=SOME_REASON, ts_accepted=1000000, ts_triggered=0, ts_last=2000000, ts_init=3000000)"  # noqa
         )
 
     def test_add_position_state_reports(self):