+++ tests/unit_tests/model/test_model_enums.py
@@ -61,8 +61,10 @@ from nautilus_trader.model.enums import TimeInForce
 from nautilus_trader.model.enums import TimeInForceParser
 from nautilus_trader.model.enums import TradingState
 from nautilus_trader.model.enums import TradingStateParser
-from nautilus_trader.model.enums import TriggerMethod
-from nautilus_trader.model.enums import TriggerMethodParser
+from nautilus_trader.model.enums import TrailingOffsetType
+from nautilus_trader.model.enums import TrailingOffsetTypeParser
+from nautilus_trader.model.enums import TriggerType
+from nautilus_trader.model.enums import TriggerTypeParser
 from nautilus_trader.model.enums import VenueStatus
 from nautilus_trader.model.enums import VenueStatusParser
 
@@ -769,6 +771,8 @@ class TestOrderType:
             [OrderType.LIMIT, "LIMIT"],
             [OrderType.STOP_MARKET, "STOP_MARKET"],
             [OrderType.STOP_LIMIT, "STOP_LIMIT"],
+            [OrderType.TRAILING_STOP_MARKET, "TRAILING_STOP_MARKET"],
+            [OrderType.TRAILING_STOP_LIMIT, "TRAILING_STOP_LIMIT"],
         ],
     )
     def test_order_type_to_str(self, enum, expected):
@@ -785,6 +789,8 @@ class TestOrderType:
             ["LIMIT", OrderType.LIMIT],
             ["STOP_MARKET", OrderType.STOP_MARKET],
             ["STOP_LIMIT", OrderType.STOP_LIMIT],
+            ["TRAILING_STOP_MARKET", OrderType.TRAILING_STOP_MARKET],
+            ["TRAILING_STOP_LIMIT", OrderType.TRAILING_STOP_LIMIT],
         ],
     )
     def test_order_type_from_str(self, string, expected):
@@ -1054,29 +1060,24 @@ class TestTradingState:
         assert expected == result
 
 
-class TestTriggerMethod:
-    def test_trigger_method_parser_given_invalid_value_raises_value_error(self):
+class TestTrailingOffsetType:
+    def test_trading_state_parser_given_invalid_value_raises_value_error(self):
         # Arrange, Act, Assert
         with pytest.raises(ValueError):
-            TriggerMethodParser.from_str_py("")
+            TrailingOffsetTypeParser.from_str_py("")
 
     @pytest.mark.parametrize(
         "enum, expected",
         [
-            [TriggerMethod.DEFAULT, "DEFAULT"],
-            [TriggerMethod.LAST, "LAST"],
-            [TriggerMethod.BID_ASK, "BID_ASK"],
-            [TriggerMethod.DOUBLE_LAST, "DOUBLE_LAST"],
-            [TriggerMethod.DOUBLE_BID_ASK, "DOUBLE_BID_ASK"],
-            [TriggerMethod.LAST_OR_BID_ASK, "LAST_OR_BID_ASK"],
-            [TriggerMethod.MID_POINT, "MID_POINT"],
-            [TriggerMethod.MARK, "MARK"],
-            [TriggerMethod.INDEX, "INDEX"],
+            [TrailingOffsetType.PRICE, "PRICE"],
+            [TrailingOffsetType.BASIS_POINTS, "BASIS_POINTS"],
+            [TrailingOffsetType.TICKS, "TICKS"],
+            [TrailingOffsetType.PRICE_TIER, "PRICE_TIER"],
         ],
     )
-    def test_trigger_method_to_str(self, enum, expected):
+    def test_trailing_offset_type_to_str(self, enum, expected):
         # Arrange, Act
-        result = TriggerMethodParser.to_str_py(enum)
+        result = TrailingOffsetTypeParser.to_str_py(enum)
 
         # Assert
         assert expected == result
@@ -1084,20 +1085,64 @@ class TestTriggerMethod:
     @pytest.mark.parametrize(
         "string, expected",
         [
-            ["DEFAULT", TriggerMethod.DEFAULT],
-            ["LAST", TriggerMethod.LAST],
-            ["BID_ASK", TriggerMethod.BID_ASK],
-            ["DOUBLE_LAST", TriggerMethod.DOUBLE_LAST],
-            ["DOUBLE_BID_ASK", TriggerMethod.DOUBLE_BID_ASK],
-            ["LAST_OR_BID_ASK", TriggerMethod.LAST_OR_BID_ASK],
-            ["MID_POINT", TriggerMethod.MID_POINT],
-            ["MARK", TriggerMethod.MARK],
-            ["INDEX", TriggerMethod.INDEX],
+            ["PRICE", TrailingOffsetType.PRICE],
+            ["BASIS_POINTS", TrailingOffsetType.BASIS_POINTS],
+            ["TICKS", TrailingOffsetType.TICKS],
+            ["PRICE_TIER", TrailingOffsetType.PRICE_TIER],
         ],
     )
-    def test_trigger_method_from_str(self, string, expected):
+    def test_trailing_offset_type_from_str(self, string, expected):
         # Arrange, Act
-        result = TriggerMethodParser.from_str_py(string)
+        result = TrailingOffsetTypeParser.from_str_py(string)
+
+        # Assert
+        assert expected == result
+
+
+class TestTriggerType:
+    def test_trigger_type_parser_given_invalid_value_raises_value_error(self):
+        # Arrange, Act, Assert
+        with pytest.raises(ValueError):
+            TriggerTypeParser.from_str_py("")
+
+    @pytest.mark.parametrize(
+        "enum, expected",
+        [
+            [TriggerType.DEFAULT, "DEFAULT"],
+            [TriggerType.LAST, "LAST"],
+            [TriggerType.BID_ASK, "BID_ASK"],
+            [TriggerType.DOUBLE_LAST, "DOUBLE_LAST"],
+            [TriggerType.DOUBLE_BID_ASK, "DOUBLE_BID_ASK"],
+            [TriggerType.LAST_OR_BID_ASK, "LAST_OR_BID_ASK"],
+            [TriggerType.MID_POINT, "MID_POINT"],
+            [TriggerType.MARK, "MARK"],
+            [TriggerType.INDEX, "INDEX"],
+        ],
+    )
+    def test_trigger_type_to_str(self, enum, expected):
+        # Arrange, Act
+        result = TriggerTypeParser.to_str_py(enum)
+
+        # Assert
+        assert expected == result
+
+    @pytest.mark.parametrize(
+        "string, expected",
+        [
+            ["DEFAULT", TriggerType.DEFAULT],
+            ["LAST", TriggerType.LAST],
+            ["BID_ASK", TriggerType.BID_ASK],
+            ["DOUBLE_LAST", TriggerType.DOUBLE_LAST],
+            ["DOUBLE_BID_ASK", TriggerType.DOUBLE_BID_ASK],
+            ["LAST_OR_BID_ASK", TriggerType.LAST_OR_BID_ASK],
+            ["MID_POINT", TriggerType.MID_POINT],
+            ["MARK", TriggerType.MARK],
+            ["INDEX", TriggerType.INDEX],
+        ],
+    )
+    def test_trigger_type_from_str(self, string, expected):
+        # Arrange, Act
+        result = TriggerTypeParser.from_str_py(string)
 
         # Assert
         assert expected == result